app/backend/requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
google-adk
google-generativeai
httpx
crawl4ai
pydantic
taskiq
playwright
yake
pylatexenc
python-jobspy
pandas
pypandoc
